@@ -1,9 +1,10 @@
 {
   "name": "hakoniwa",
-  "version": "0.1.3",
+  "version": "0.1.4-alpha.1",
   "description": "miniature wonder toolbox for front-end development.",
-  "main": "index.browser.js",
-  "types": "index.browser.d.ts",
+  "exports": "./index.browser.js",
+  "types": "./index.browser.d.ts",
+  "type": "commonjs",
   "files": [
     "src",
     "spec",
@@ -16,7 +17,7 @@
     "scripts"
   ],
   "scripts": {
-    "prepare": "npm run build",
+    "prepare": "npm run clean && npm run build",
     "build": "tsc",
     "test": "ts-node ./spec/whistle.spec.ts",
     "watch": "tsc -w",
@@ -38,40 +39,40 @@
   },
   "homepage": "https://github.com/lonelyhentai/hakoniwa#readme",
   "devDependencies": {
-    "@types/bluebird": "^3.5.33",
-    "@types/hapi__joi": "^17.1.4",
-    "@types/lodash": "^4.14.160",
-    "@types/node": "^14.14.34",
-    "@types/node-fetch": "^2.5.7",
-    "@typescript-eslint/eslint-plugin": "^4.17.0",
-    "@typescript-eslint/parser": "^4.17.0",
-    "commitizen": "^4.2.1",
+    "@types/bluebird": "^3.5.36",
+    "@types/lodash": "^4.14.176",
+    "@types/node": "^16.11.6",
+    "@types/node-fetch": "^2.0.3",
+    "@typescript-eslint/eslint-plugin": "^5.2.0",
+    "@typescript-eslint/parser": "^5.2.0",
+    "commitizen": "^4.2.4",
     "cz-conventional-changelog": "^3.3.0",
-    "eslint": "^7.7.0",
-    "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.20.6",
-    "prettier": "^2.1.1",
-    "ts-node": "^9.0.0",
-    "tsc-watch": "^4.2.9",
-    "tsconfig-paths": "^3.9.0",
-    "typescript": "^4.2.3"
+    "eslint": "^8.1.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.26.1",
+    "prettier": "^2.4.1",
+    "ts-node": "^10.4.0",
+    "tsc-watch": "^4.5.0",
+    "tsconfig-paths": "^3.11.0",
+    "tslib": "^2.3.1",
+    "typescript": "^4.4.4"
   },
   "dependencies": {
     "@cypress/skip-test": "^2.6.0",
-    "@hapi/joi": "^17.1.1",
     "chai-colors": "^1.0.1",
     "cross-env": "^7.0.2",
-    "cypress": "^6.6.0",
-    "cypress-localstorage-commands": "^1.4.1",
+    "cypress": "^8.7.0",
+    "cypress-localstorage-commands": "^1.5.0",
     "cypress-promise": "^1.1.0",
-    "cypress-skip-and-only-ui": "^1.2.7",
+    "cypress-skip-and-only-ui": "^1.2.11",
     "cypress-wait-until": "^1.7.1",
-    "json5": "^2.1.3",
-    "lodash": "^4.17.20",
-    "node-fetch": "^2.6.1",
+    "joi": "^17.4.2",
+    "json5": "^2.2.0",
+    "lodash": "^4.17.21",
+    "node-fetch": "^2.6.5",
     "queue-typescript": "^1.0.1",
-    "whistle": "^2.6.7"
+    "whistle": "^2.8.0"
   },
   "bin": {
     "hakoniwa-cli": "./scripts/cli.js"
